pipeline/requirements.txt
transformers==4.38.2
accelerate==1.0.0
evaluate
av
librosa==0.10.1
soundfile
resampy
pytorchvideo
easyocr
pandas
numpy==1.26.4
scipy
scikit-learn
matplotlib
seaborn
datasets
pymongo[srv]==3.12
kafka-python
Pillow
huggingface-hub
safetensors
tokenizers
opencv-python
imageio
scikit-image
webdriver-manager
selenium
lxml
joblib
tqdm
colorama
psutil
certifi
browser_cookie3
bs4
requests
pygments
wordcloud
streamlit_autorefresh
wcwidth
